backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# AWS
boto3>=1.34.0

# Transcription
faster-whisper>=1.0.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0
aiofiles>=23.2.0
python-multipart>=0.0.6

# Google API (for Meet)
google-auth>=2.27.0
google-api-python-client>=2.111.0
